Cleaned up the user-roster area, moved the selection card to its own component
