+++ react-frontend/src/Components/UserRoster/PlayerOffcanvas.jsx
@@ -12,6 +12,7 @@ import { PlayersContext } from "../../Contexts/PlayersContext"
 import { EventConfig } from "../../Contexts/EventConfig"
 import HandleDBTransaction from "../../Functions/HandleDBTransaction";
 import BASE_URL from "../../assets/Files/BASE_URL";
+import PlayerSelectionCard from "./PlayerSelectionCard";
 
 
 
@@ -19,8 +20,6 @@ const PlayerOffcanvas = ({ show, cardName, setShow, i }) => {
   const { playersContext, setPlayersContext } = useContext(PlayersContext)
   const { currentUser, setCurrentUser } = useContext(CurrentUser)
   const { eventConfig, setEventConfig } = useContext(EventConfig)
-  let [message, setMessage] = useState(null)
-  let [alert, setAlert] = useState(false)
 
   if (playersContext &&  currentUser && eventConfig) {
     const { dataSettings } = eventConfig
@@ -28,7 +27,6 @@ const PlayerOffcanvas = ({ show, cardName, setShow, i }) => {
     const { players } = playersContext
     const { user_id, roster } = currentUser
     const { past_champ, usa, intl, wild_card1, wild_card2, wild_card3 } = roster
-    // console.log(currentUser)
 
     let playerList = []
     let playerRoster = [past_champ, usa, intl, wild_card1, wild_card2, wild_card3]
@@ -55,86 +53,13 @@ const PlayerOffcanvas = ({ show, cardName, setShow, i }) => {
 
     const handleClose = () => setShow(false);
 
-    const handleSelect =  async (e, id) => {
-      const { roster } = currentUser
-      const { year } = roster
-
-      // if a roster record does not exist 
-    if (!year) {
-      // create record in DB
-      let path = BASE_URL + 'roster/new'
-      let payload = {
-        year:  tournamentYear,
-        [key]: id,
-        user_id: user_id
-      }
-
-      try {
-        let insertResponse = await HandleDBTransaction(path, 'POST', payload)
-        let { status } = insertResponse
-
-        if (status === 201) {
-          location.reload()
-        } else {
-          let data = await insertResponse.json()
-          setMessage(data.msg)
-          setAlert(true)
-        }
-      } catch (error) {
-        setMessage('Roster not saved')
-        setAlert(true)
-      }
-    } else {
-      // Record already exists - UPDATE
-      let path = BASE_URL + 'roster/' + user_id
-      let payload = {
-        [key]: id      
-      }
-
-      try {
-        let updateResponse = await HandleDBTransaction(path, 'PUT', payload)
-        let { status } = updateResponse
-
-        if (status === 201) {
-          location.reload()
-        } else {
-          let data = await updateResponse.json()
-          setMessage(data.msg)
-          setAlert(true)
-        }
-      } catch (error) {
-        setMessage('Roster not saved')
-        setAlert(true)
-      }
-    }
-      
-      
-      // save player to the DB
-      console.log(currentUser) 
-      // update the current user
-      // close offcanvas
-      // reload page?
-      // location.reload()
-
-    } 
 
     const playerSelection = playerList.map(player => {
-      const { id, first_name , last_name } = player
+      const { id } = player
       const picture = `https://images.masters.com/players/${tournamentYear}/240x240/${id}.jpg`
 
       return (
-        < Container fluid key={`selection-card-${i}-${id}`}>
-          <Card className={`m-1 p-1 text-center${id == roster[key] ? ' border border-success shadow' : (
-            playerRoster.includes(Number(id)) ? ' border border-danger shadow' : ''
-          )}`} >
-            <Image src={picture} className=' mx-auto border rounded-circle roster-img' />
-            <Card.Body>
-              <Card.Title>{first_name} {last_name} {player.amateur ? '(A)' : null}</Card.Title>
-              <Button variant="success" onClick={(e) => handleSelect(e, id)} disabled={playerRoster.includes(Number(id))}>Select Player</Button>
-              <Alert variant='danger' show={alert}>{message} </Alert>
-            </Card.Body>
-          </Card>
-        </Container>
+        <PlayerSelectionCard player={player}picture={picture} disable={playerRoster.includes(Number(id))} current={id == roster[key]} currentUser={currentUser} tournamentYear={tournamentYear} column={key} key={`selection-card-${i}-${id}`}/>
       )
       
     })
